requirements.txt
apprise==1.9.4
praw==7.8.1
python-dotenv==1.1.1
tenacity==9.1.2
Markdown==3.9
PyYAML==6.0.2
click==8.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
oauthlib==3.3.1
requests==2.32.5
requests-oauthlib==2.0.0
urllib3==2.5.0
websocket-client==1.8.0
update-checker==0.18.0
